+++ graphics/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ## Unreleased
 
+### Changed
+- (Breaking)Updated Comsic_text to 0.15.0
+- Included Alignment Option for set_text, measure_string and measure_glyphs.
+
 ## 0.28.0 (7. October, 2025)
 
 ### Changed